refactor: [UXP-3578] Cache factory name and parameters
